Adds client intro rendering.
