+++ packages/reactotron-app/App/Streaming/Commands/index.js
@@ -2,6 +2,7 @@ import LogCommand from './LogCommand'
 import UnknownCommand from './UnknownCommand'
 import StateActionCompleteCommand from './StateActionCompleteCommand'
 import ApiResponseCommand from './ApiResponseCommand'
+import ClientIntroCommand from './ClientIntroCommand'
 
 export default command => {
   const { type } = command
@@ -11,6 +12,7 @@ export default command => {
     case 'log': return LogCommand
     case 'state.action.complete': return StateActionCompleteCommand
     case 'api.response': return ApiResponseCommand
+    case 'client.intro': return ClientIntroCommand
     default: return UnknownCommand
   }
 }
